Organized item classes
Pulled item classes out of Items into their own seperate files.
