+++ src/main/java/iamtheissue/myutils/Main.java
@@ -3,6 +3,7 @@ package iamtheissue.myutils;
 import iamtheissue.myutils.init.Blocks;
 import iamtheissue.myutils.init.Items;
 import iamtheissue.myutils.init.Recipes;
+import iamtheissue.myutils.items.FuelHandler;
 import iamtheissue.myutils.proxy.CommonProxy;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.Loader;
@@ -28,7 +29,7 @@ public class Main
 		Items.init();
 		Items.register();
 		Items.preInit();
-		GameRegistry.registerFuelHandler(new Items.FuelHandler());
+		GameRegistry.registerFuelHandler(new FuelHandler());
 		
 	}
