+++ src/main/resources/assets/mu/lang/en_US.lang
@@ -8,8 +8,8 @@ item.raw_spider_leg.name=Raw Spider Leg
 item.cooked_spider_leg.name=Cooked Spider Leg
 item.coal_pellet.name=Coal Pellet
 item.charcoal_pellet.name=Charcoal Pellet
-item.phial_goo_half.name=Half a Phial of Tainted Stuff
-item.phial_goo_full.name=Phial of Tainted Stuff
+item.phial_goo_0.name=Half a Phial of Tainted Stuff
+item.phial_goo_1.name=Phial of Tainted Stuff
 item.cooked_spider_eye.name=Cooked Spider Eye
 tile.lamp.name=Smooth Glowstone
 tile.glowstone_block.name=Glowstone Block
