+++ src/main/resources/assets/mu/models/item/phial_goo_0.json
@@ -1,7 +1,7 @@
 {
     "parent": "builtin/generated",
     "textures": {
-        "layer0": "mu:items/phial_goo_half"
+        "layer0": "mu:items/phial_goo_0"
     },
     "display": {
         "thirdperson": {
