+++ src/main/resources/assets/mu/textures/items/phial_goo_0.png
