+++ src/main/resources/mcmod.info
@@ -3,7 +3,7 @@
   "modid": "mu",
   "name": "My Utils",
   "description": "Adds some random Stuff that i found the game lacked.",
-  "version": "1.2",
+  "version": "1.3",
   "mcversion": "1.8.9",
   "url": "",
   "updateUrl": "",
